# swagger.yaml
openapi: 3.0.3
info:
  title: AdoptMe API
  description: |
    API para la plataforma de adopción de mascotas AdoptMe.
    Permite gestionar usuarios, mascotas y adopciones.
  version: 1.0.0
  contact:
    name: Soporte
    email: soporte@adoptme.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081
    description: Servidor de desarrollo
  - url: https://api.adoptme.com
    description: Servidor de producción
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Ingrese el token JWT en el formato: Bearer <token>
        
        Ejemplo: `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
  schemas:
...
    PetInput:
      type: object
      required: [name, specie, birthDate]
      properties:
        name:
          type: string
          example: Firulais
          description: Nombre de la mascota
        specie:
          type: string
          example: Perro
          description: Especie de la mascota
          enum: [Perro, Gato, Conejo, Otro]
        birthDate:
          type: string
          format: date
          example: '2020-01-01'
          description: Fecha de nacimiento (YYYY-MM-DD)
        description:
          type: string
          example: Muy juguetón y cariñoso
          description: Descripción de la mascota
        image:
          type: string
          format: uri
          example: https://example.com/pet.jpg
          description: URL de la imagen de la mascota
    Pet:
      allOf:
        - $ref: '#/components/schemas/PetInput'
        - type: object
          properties:
            _id:
              type: string
              example: 60f7b1b3b3f3b3b3b3b3b3b3b
            owner:
              type: string
              description: ID del dueño
              example: 60f7b1b3b3f3b3b3b3b3b3b3a
            adopted:
              type: boolean
              default: false
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    Adoption:
      type: object
      properties:
        _id:
          type: string
          example: 60f7b1b3b3f3b3b3b3b3b3b3c
        owner:
          type: string
          description: ID del usuario dueño
          example: 60f7b1b3b3f3b3b3b3b3b3b3a
        pet:
          type: string
          description: ID de la mascota
          example: 60f7b1b3b3f3b3b3b3b3b3b3b
        status:
          type: string
          enum: [pending, approved, rejected]
          default: pending
        notes:
          type: string
          example: El adoptante parece ser una buena opción
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        status:
          type: string
          example: error
        error:
          type: string
          example: Mensaje de error descriptivo
    ValidationError:
      type: object
      properties:
        status:
          type: string
          example: error
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: Debe ser un correo electrónico válido
...
security:
  - bearerAuth: []
paths:
...
  /api/pets:
    get:
      tags: [Pets]
      summary: Obtener todas las mascotas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de mascotas
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  payload:
                    type: array
                    items:
                      $ref: '#/components/schemas/Pet'
    
    post:
      tags: [Pets]
      summary: Crear una nueva mascota
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PetInput'
      responses:
        '201':
          description: Mascota creada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  payload:
                    $ref: '#/components/schemas/Pet'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /api/pets/{pid}:
    put:
      tags: [Pets]
      summary: Actualizar una mascota
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: pid
          required: true
          schema:
            type: string
          description: ID de la mascota a actualizar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PetInput'
      responses:
        '200':
          description: Mascota actualizada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  payload:
                    $ref: '#/components/schemas/Pet'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '404':
          description: Mascota no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Pets]
      summary: Eliminar una mascota
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: pid
          required: true
          schema:
            type: string
          description: ID de la mascota a eliminar
      responses:
        '200':
          description: Mascota eliminada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Mascota eliminada exitosamente
        '404':
          description: Mascota no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/adoptions:
    get:
      tags: [Adoptions]
      summary: Obtener todas las adopciones
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de adopciones
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  payload:
                    type: array
                    items:
                      $ref: '#/components/schemas/Adoption'
  /api/adoptions/{aid}:
    get:
      tags: [Adoptions]
      summary: Obtener una adopción por ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: aid
          required: true
          schema:
            type: string
          description: ID de la adopción a buscar
      responses:
        '200':
          description: Adopción encontrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  payload:
                    $ref: '#/components/schemas/Adoption'
        '404':
          description: Adopción no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/adoptions/{uid}/{pid}:
    post:
      tags: [Adoptions]
      summary: Crear una nueva solicitud de adopción
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: uid
          required: true
          schema:
            type: string
          description: ID del usuario que solicita la adopción
        - in: path
          name: pid
          required: true
          schema:
            type: string
          description: ID de la mascota a adoptar
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
                  description: Notas adicionales para la solicitud de adopción
      responses:
        '201':
          description: Solicitud de adopción creada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  payload:
                    $ref: '#/components/schemas/Adoption'
        '400':
          description: La mascota ya ha sido adoptada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Usuario o mascota no encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'